Tests:  change remote procedure numbers to Enums. Rename test to Client Should Call Remote Procedures.  Comment out java serial number leftover.

+++ src/rpc/oncrpc.mstest/Tcp/OncRpcTcpTestClient.cs
@@ -113,7 +113,7 @@ public class OncRpcTcpTestClient : IDisposable
     {
         VoidXdrCodec args = VoidXdrCodec.VoidXdrCodecInstance;
         VoidXdrCodec result = VoidXdrCodec.VoidXdrCodecInstance;
-        this.Call( this.Version == 1 ? RemoteProceduresVersion1.Nop : RemoteProceduresVersion2.Nop, this.Version, args, result );
+        this.Call( this.Version == 1 ? ( int ) RemoteProceduresVersion1.Nop : ( int ) RemoteProceduresVersion2.Nop, this.Version, args, result );
     }
 
     /// <summary>
@@ -125,7 +125,7 @@ public class OncRpcTcpTestClient : IDisposable
     {
         StringXdrCodec args = new( arg1 );
         StringXdrCodec result = new();
-        this.Call( RemoteProceduresVersion1.Echo, this.Version, args, result );
+        this.Call( ( int ) RemoteProceduresVersion1.Echo, this.Version, args, result );
         return result.Value;
     }
 
@@ -138,7 +138,7 @@ public class OncRpcTcpTestClient : IDisposable
     public virtual string CallRemoteProcedureConcatenateInputParameters( StringVectorCodec arg1 )
     {
         StringXdrCodec result = new();
-        this.Call( RemoteProceduresVersion1.ConcatenateInputParameters, this.Version, arg1, result );
+        this.Call( ( int ) RemoteProceduresVersion1.ConcatenateInputParameters, this.Version, arg1, result );
         return result.Value;
     }
 
@@ -151,7 +151,7 @@ public class OncRpcTcpTestClient : IDisposable
     {
         IntXdrCodec args = new( arg1 );
         BooleanXdrCodec result = new();
-        this.Call( RemoteProceduresVersion1.CompareInputToFoo, this.Version, args, result );
+        this.Call( (int) RemoteProceduresVersion1.CompareInputToFoo, this.Version, args, result );
         return result.Value;
     }
 
@@ -163,7 +163,7 @@ public class OncRpcTcpTestClient : IDisposable
     {
         VoidXdrCodec args = VoidXdrCodec.VoidXdrCodecInstance;
         IntXdrCodec result = new();
-        this.Call( RemoteProceduresVersion1.ReturnEnumFooValue, this.Version, args, result );
+        this.Call( (int) RemoteProceduresVersion1.ReturnEnumFooValue, this.Version, args, result );
         return result.Value;
     }
 
@@ -178,7 +178,7 @@ public class OncRpcTcpTestClient : IDisposable
     public virtual LinkedListCodec CallRemoteProcedureBuildLinkedList( LinkedListCodec arg1 )
     {
         LinkedListCodec result = new();
-        this.Call( RemoteProceduresVersion1.BuildLinkedList, this.Version, arg1, result );
+        this.Call( (int) RemoteProceduresVersion1.BuildLinkedList, this.Version, arg1, result );
         return result;
     }
 
@@ -192,7 +192,7 @@ public class OncRpcTcpTestClient : IDisposable
     {
         VoidXdrCodec args = VoidXdrCodec.VoidXdrCodecInstance;
         SomeResultCodec result = new();
-        this.Call( RemoteProceduresVersion1.RemoteProcedureReadSomeResult, this.Version, args, result );
+        this.Call( (int) RemoteProceduresVersion1.RemoteProcedureReadSomeResult, this.Version, args, result );
         return result;
     }
 
@@ -207,7 +207,7 @@ public class OncRpcTcpTestClient : IDisposable
             Arg2 = arg2
         };
         StringXdrCodec result = new();
-        this.Call( RemoteProceduresVersion2.ConcatenateTwoValues, RpcProgramConstants.Version2, args, result );
+        this.Call( (int) RemoteProceduresVersion2.ConcatenateTwoValues, RpcProgramConstants.Version2, args, result );
         return result.Value;
     }
 
@@ -229,7 +229,7 @@ public class OncRpcTcpTestClient : IDisposable
             Three = three
         };
         StringXdrCodec result = new();
-        this.Call( RemoteProceduresVersion2.ConcatenateThreeItems, RpcProgramConstants.Version2, args, result );
+        this.Call( ( int ) RemoteProceduresVersion2.ConcatenateThreeItems, RpcProgramConstants.Version2, args, result );
         return result.Value;
     }
 
@@ -242,7 +242,7 @@ public class OncRpcTcpTestClient : IDisposable
     {
         IntXdrCodec args = new( foo );
         StringXdrCodec result = new();
-        this.Call( RemoteProceduresVersion2.ReturnYouAreFooValue, RpcProgramConstants.Version2, args, result );
+        this.Call( ( int ) RemoteProceduresVersion2.ReturnYouAreFooValue, RpcProgramConstants.Version2, args, result );
         return result.Value;
     }
 
@@ -263,7 +263,7 @@ public class OncRpcTcpTestClient : IDisposable
             List2 = list2
         };
         LinkedListCodec result = new();
-        this.Call( RemoteProceduresVersion2.LinkListItems, RpcProgramConstants.Version2, args, result );
+        this.Call( (int) RemoteProceduresVersion2.LinkListItems, RpcProgramConstants.Version2, args, result );
         return result;
     }
 
@@ -281,7 +281,7 @@ public class OncRpcTcpTestClient : IDisposable
             D = d
         };
         VoidXdrCodec result = VoidXdrCodec.VoidXdrCodecInstance;
-        this.Call( RemoteProceduresVersion2.ProcessFourArguments, RpcProgramConstants.Version2, args, result );
+        this.Call( (int) RemoteProceduresVersion2.ProcessFourArguments, RpcProgramConstants.Version2, args, result );
     }
